[awk] Support HW and SW drawing on Kitkat
SW rendering on Kitkat already uses AwPixelInfo version 3.

It also claims to use SkCanvasState version 1 but the class layout
 between Kitkat and upstream differs.
* Kitkat is missing int32_t SkCanvasState.alignmentPadding
* Kitkat has size_t SkCanvasState.rester.rowBytes instead of uint64_t

See also
* https://android.googlesource.com/platform/external/chromium_org/+/refs/heads/kitkat-release/android_webview/public/browser/draw_sw.h
* https://source.chromium.org/chromium/_/skia/skia.git/+/352c2181d15ed053c3b759f08ff1f51d50e2d3bb
* https://chromium.googlesource.com/chromium/src/+/HEAD/base/memory/raw_ptr.md
* https://chromium-review.googlesource.com/c/chromium/src/+/3607476

+++ android_webview/glue/java/src/com/android/webview/chromium/GraphicsUtilsCompat.java
@@ -24,7 +24,7 @@ final class GraphicsUtilsCompat {
         }
     }
 
-    private static long getDrawSWFunctionTableInt() {
+    private static int getDrawSWFunctionTableInt() {
         try {
             return (int) getGraphicsUtilsClass()
                     .getMethod("getDrawSWFunctionTable")
@@ -34,7 +34,7 @@ final class GraphicsUtilsCompat {
         }
     }
 
-    private static long getDrawGLFunctionTableInt() {
+    private static int getDrawGLFunctionTableInt() {
         try {
             return (int) getGraphicsUtilsClass()
                     .getMethod("getDrawGLFunctionTable")
